+++ tide/core/thumbnail/ThumbnailProvider.cpp
@@ -45,7 +45,6 @@
 #include <QCache>
 #include <QDateTime>
 #include <QFileInfo>
-#include <QImageReader>
 
 #include <cassert>